core: move AccessWitness from core/types to core/state
