+++ core/state_processor.go
@@ -96,7 +96,7 @@ func applyTransaction(msg types.Message, config *params.ChainConfig, bc ChainCon
 	// Create a new context to be used in the EVM environment.
 	txContext := NewEVMTxContext(msg)
 	if config.IsCancun(blockNumber) {
-		txContext.Accesses = types.NewAccessWitness()
+		txContext.Accesses = state.NewAccessWitness()
 	}
 	evm.Reset(txContext, statedb)
