+++ core/vm/interface.go
@@ -20,6 +20,7 @@ import (
 	"math/big"
 
 	"github.com/ethereum/go-ethereum/common"
+	"github.com/ethereum/go-ethereum/core/state"
 	"github.com/ethereum/go-ethereum/core/types"
 )
 
@@ -78,8 +79,8 @@ type StateDB interface {
 
 	ForEachStorage(common.Address, func(common.Hash, common.Hash) bool) error
 
-	Witness() *types.AccessWitness
-	SetWitness(*types.AccessWitness)
+	Witness() *state.AccessWitness
+	SetWitness(*state.AccessWitness)
 }
 
 // CallContext provides a basic interface for the EVM calling conventions. The EVM
